+++ city_metrix/__init__.py
@@ -1,5 +1,6 @@
 import os
 import warnings
+from pathlib import Path
 
 import boto3
 import ee
@@ -27,15 +28,17 @@ else:
     ee.Initialize(opt_url="https://earthengine-highvolume.googleapis.com")
 
 # initialize aws
+credentials_file_path = Path(os.path.join(Path.home(),'.aws', 'credentials'))
 if (
     "AWS_ACCESS_KEY_ID" in os.environ
     and "AWS_SECRET_ACCESS_KEY" in os.environ
 ):
-    s3_client = boto3.client('s3')
-else:
-    session = boto3.Session(profile_name='cities-data-dev')
+    s3_client = boto3.client('s3', region_name='us-east-1')
+elif credentials_file_path.exists():
+    session = boto3.Session(profile_name='cities-data-dev', region_name='us-east-1')
     s3_client = session.client('s3')
-
+else:
+    Exception("Could not establish AWS credentials")
 
 # set for AWS requests
 os.environ["AWS_REQUEST_PAYER"] = "requester"