+++ city_metrix/constants.py
@@ -18,8 +18,8 @@ RW_DASHBOARD_LAYER_S3_BUCKET_URI = 's3://cities-test-sandbox' # 's3://wri-cities
 RW_DASHBOARD_METRIC_S3_BUCKET_URI = 's3://wri-cities-indicators'
 RW_TESTING_S3_BUCKET_URI = 's3://cities-test-sandbox'
 
-home_directory = os.path.join(Path.home(), 'CIF_layer_repository')
-LOCAL_REPO_URI = f'file://{home_directory}'
+repo_directory = os.path.join(Path.home(), 'CIF_layer_repository')
+LOCAL_REPO_URI = f'file://{repo_directory}'
 
 CITIES_DATA_API_URL = "cities-data-api.wri.org"