+++ tests/resources/metrics_dumps/test_write_all_metrics_by_city.py
@@ -11,6 +11,7 @@ from ..tools import prep_output_path, cleanup_cache_files
 PRESERVE_RESULTS_ON_S3 = True # True - Default for check-in
 PRESERVE_RESULTS_ON_OS = False # False - Default for check-in
 FORCE_DATA_REFRESH = False # False - Default for check-in
+
 SLOW_TEST_TIMEOUT_SECONDS = 60 # 1 minute
 
 PROCESSING_CITY = GEOEXTENT_TERESINA_WGS84